# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "健身计划管理系统API",
        "title": "Flow Link Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/training/records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取训练记录列表，支持分页和筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "训练记录"
                ],
                "summary": "获取训练记录列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "开始日期",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "结束日期",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "计划ID",
                        "name": "planId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.PaginatedData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的训练记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "训练记录"
                ],
                "summary": "创建训练记录",
                "parameters": [
                    {
                        "description": "训练记录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateTrainingRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TrainingRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/training/records/{recordId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取训练记录详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "训练记录"
                ],
                "summary": "获取训练记录详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "recordId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TrainingRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "记录ID不能为空",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "训练记录不存在",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新指定ID的训练记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "训练记录"
                ],
                "summary": "更新训练记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "recordId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "训练记录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateTrainingRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除指定ID的训练记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "训练记录"
                ],
                "summary": "删除训练记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "recordId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "记录ID不能为空",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateTrainingRecordRequest": {
            "type": "object",
            "required": [
                "duration",
                "endTime",
                "exercises",
                "startTime",
                "title"
            ],
            "properties": {
                "caloriesBurned": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "endTime": {
                    "description": "YYYY-MM-DD HH:mm:ss",
                    "type": "string"
                },
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Exercise"
                    }
                },
                "mood": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "planId": {
                    "type": "integer"
                },
                "startTime": {
                    "description": "YYYY-MM-DD HH:mm:ss",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totalSets": {
                    "type": "integer"
                },
                "totalWeight": {
                    "type": "number"
                }
            }
        },
...
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Exercise": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "训练时长(分钟)",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "muscleGroup": {
                    "description": "目标肌群",
                    "type": "string"
                },
                "name": {
                    "description": "项目名称",
                    "type": "string"
                },
                "notes": {
                    "description": "备注",
                    "type": "string"
                },
                "reps": {
                    "description": "次数",
                    "type": "integer"
                },
                "restTime": {
                    "description": "休息时间(秒)",
                    "type": "integer"
                },
                "sets": {
                    "description": "组数",
                    "type": "integer"
                },
                "weight": {
                    "description": "重量(kg)",
                    "type": "number"
                }
            }
        },
...
        "domain.PaginatedData": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "plans": {
                    "description": "用于计划"
                },
                "records": {
                    "description": "用于训练记录"
                },
                "templates": {
                    "description": "用于模板"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.TrainingRecord": {
            "type": "object",
            "properties": {
                "caloriesBurned": {
                    "description": "消耗卡路里",
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "duration": {
                    "description": "总时长(分钟)",
                    "type": "integer"
                },
                "endTime": {
                    "description": "结束时间 YYYY-MM-DD HH:mm:ss",
                    "type": "string"
                },
                "exercises": {
                    "description": "训练项目列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Exercise"
                    }
                },
                "id": {
                    "type": "string"
                },
                "mood": {
                    "description": "训练状态(优秀/良好/一般/疲劳)",
                    "type": "string"
                },
                "notes": {
                    "description": "训练备注",
                    "type": "string"
                },
                "planId": {
                    "description": "关联计划ID(0表示无计划)",
                    "type": "integer"
                },
                "startTime": {
                    "description": "开始时间 YYYY-MM-DD HH:mm:ss",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totalSets": {
                    "description": "总组数",
                    "type": "integer"
                },
                "totalWeight": {
                    "description": "总重量(kg)",
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "domain.UpdateTrainingRecordRequest": {
            "type": "object",
            "properties": {
                "caloriesBurned": {
                    "type": "integer"
                },
                "duration": {
                    "type": "integer"
                },
                "endTime": {
                    "description": "YYYY-MM-DD HH:mm:ss",
                    "type": "string"
                },
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Exercise"
                    }
                },
                "mood": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "planId": {
                    "type": "integer"
                },
                "startTime": {
                    "description": "YYYY-MM-DD HH:mm:ss",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "totalSets": {
                    "type": "integer"
                },
                "totalWeight": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}